+++ lib/processors/Processor.ts
@@ -10,5 +10,7 @@ export abstract class Processor {
    * @param inputFrameBuffer - The source of the input frame to process.
    * @param outputFrameBuffer - The output frame buffer to use to draw the processed frame.
    */
-  abstract processFrame(inputFrameBuffer: OffscreenCanvas, outputFrameBuffer: HTMLCanvasElement): Promise<void> | void;
+  abstract processFrame(
+    inputFrameBuffer: OffscreenCanvas | HTMLCanvasElement | HTMLVideoElement,
+    outputFrameBuffer: HTMLCanvasElement): Promise<void> | void;
 }