+++ tests/integration/spec/GaussianBlurBackgroundProcessor.ts
@@ -1,9 +1,10 @@
 import * as assert from 'assert';
 import { GaussianBlurBackgroundProcessor, GaussianBlurBackgroundProcessorOptions } from '../../../lib/index';
-import { compareImages, getImageFromCanvas, loadImage } from '../util';
+import { compareImages, getImageFromCanvas, loadImage, pause } from '../util';
+import { Pipeline } from '../../../lib/index';
 
 describe('GaussianBlurBackgroundProcessor', function() {
-  this.timeout(30000);
+  this.timeout(60000);
 
   it('should create instance', async () => {
     const options: GaussianBlurBackgroundProcessorOptions = {
@@ -18,6 +19,7 @@ describe('GaussianBlurBackgroundProcessor', function() {
 
   describe('should render correct output', async () => {
     const runTest = async (options?: any) => {
+      options = options || {};
       const inputCanvas = new OffscreenCanvas(1, 1);
       const outputCanvas = document.createElement('canvas');
       const inputImage = await loadImage('/images/input/input_person.jpg');
@@ -30,26 +32,58 @@ describe('GaussianBlurBackgroundProcessor', function() {
 
       const processor = new GaussianBlurBackgroundProcessor({ ...options, assetsPath: '/assets' });
       await processor.loadModel();
-      await processor.processFrame(inputCanvas, outputCanvas);
 
+      const processCount = options.processCount || 1;
+      await new Promise(resolve => {
+        let counter = 0;
+        const render = () => {
+          if (counter >= processCount) {
+            return resolve(null);
+          }
+          processor.processFrame(inputCanvas, outputCanvas).then(() => {
+            counter++;
+            setTimeout(render, 0);
+          });
+        };
+        render();
+      });
       return getImageFromCanvas(outputCanvas);
     };
-    [{
-      name: 'using default configuration',
-      expectedImageName: 'blur_default',
-    },{
-      name: 'using non default blurFilterRadius',
-      expectedImageName: 'blur_filter_radius_30',
-      options: { blurFilterRadius: 30 },
-    },{
-      name: 'using non default maskBlurRadius',
-      expectedImageName: 'blur_mask_blur_100',
-      options: { maskBlurRadius: 100 },
-    }].forEach(({ name, options, expectedImageName }) => {
-      it(`when ${name}`, async () => {
-        const outputImageResult = await runTest(options);
-        const expectedOutputImage = await loadImage(`/images/output/${expectedImageName}.png`);
-        await compareImages(outputImageResult, expectedOutputImage);
+
+    describe('with Canvas2D pipeline', () => {
+      [{
+        name: 'using default configuration',
+        expectedImageName: 'blur_default',
+      },{
+        name: 'using non default blurFilterRadius',
+        expectedImageName: 'blur_filter_radius_30',
+        options: { blurFilterRadius: 30 },
+      },{
+        name: 'using non default maskBlurRadius',
+        expectedImageName: 'blur_mask_blur_100',
+        options: { maskBlurRadius: 100 },
+      }].forEach(({ name, options, expectedImageName }) => {
+        it(`when ${name}`, async () => {
+          const outputImageResult = await runTest({ ...options, pipeline: Pipeline.Canvas2D });
+          const expectedOutputImage = await loadImage(`/images/output/${expectedImageName}.png`);
+          await compareImages(outputImageResult, expectedOutputImage);
+        });
+      });
+    });
+
+    it('with WebGL2 pipeline', async () => {
+      // It takes at about 5 process frame calls to render the GPU contents out to the canvas
+      const outputImageResult = await runTest({ processCount: 5 });
+      const expectedOutputImage = await loadImage('/images/output/blur_webgl.png');
+      const Rembrandt = (window as any).Rembrandt;
+
+      // WebGL renders pixels with slight variation per run. Let's put some thresholds
+      // to account for those. See http://rembrandtjs.com for more information
+      await compareImages(outputImageResult, expectedOutputImage, {
+        thresholdType: Rembrandt.THRESHOLD_PERCENT,
+        maxThreshold: 0.01,
+        maxDelta: 20,
+        maxOffset: 0,
       });
     });
   });